fixed a bunch of stuff, paint109 is looking close

+++ Wrap109.py
@@ -10,11 +10,11 @@ class Wrap109:
 
     # assumes CV_32F mat
     def get_frame(self, src, threshold):
-        temp = src.astype('uint8').copy()
+        temp = src.copy()
         temp = fof.thresh(temp, threshold)
         temp = fof.contour(temp, .01, 9, 1.5, (255, 255, 255))
 
-        temp = self.trails.get_frame(temp.astype('float32'), 9.0, 0.8)
+        temp = self.trails.get_frame(temp, 9.0, 0.8)
 
         temp = 2 * temp + .6 * src
         return temp
